fix algorithm dependency problems
A bunch of different headers required detail::all_of, and it was unclear where to put it.
The solution is to create a new header containing only that algorithm, which reduces the amount of <algorithm> dependencies.
This technique can also be applied for some other menial dependencies, such as including <algorithm> solely for std::min and std::max.